Fixes for initial implementation UserController
